+++ bluetooth/src/commonMain/kotlin/ScanningStateRepo.kt
@@ -1,0 +1,17 @@
+package com.splendo.kaluga.bluetooth
+
+import com.splendo.kaluga.bluetooth.device.Device
+import com.splendo.kaluga.log.LogLevel
+import com.splendo.kaluga.log.logger
+import com.splendo.kaluga.permissions.Permissions
+import com.splendo.kaluga.permissions.Support
+import com.splendo.kaluga.state.State
+import com.splendo.kaluga.state.StateRepo
+import com.splendo.kaluga.state.StateRepoAccesor
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
+
+
+
+
+
