+++ bluetooth/src/commonMain/kotlin/device/Device.kt
@@ -1,0 +1,14 @@
+package com.splendo.kaluga.bluetooth.device
+
+
+expect class Identifier
+
+interface BaseDevice {
+    val name: String?
+    val identifier: Identifier
+    val advertisementData: AdvertisementData
+}
+
+expect class Device : BaseDevice {
+
+}
