+++ bluetooth/src/commonMain/kotlin/device/DeviceState.kt
@@ -1,0 +1,2 @@
+package com.splendo.kaluga.bluetooth.device
+
